refactor: extract class for making map and apply singleton pattern
