+++ src/main/kotlin/bridge/view/OutputView.kt
@@ -14,40 +14,8 @@ class OutputView {
      *
      * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
      */
-    fun printMap(blocks: List<PlayerBlock>) {
-        val up = StringBuilder("")
-        val down = StringBuilder("")
-        appendMark(up, down, MapMark.OPEN)
-        blocks.forEachIndexed { index, playerBlock ->
-            appendBlock(playerBlock, up, down)
-            if (index < blocks.size - 1)
-                appendMark(up, down, MapMark.BOUNDARY)
-        }
-        appendMark(up, down, MapMark.CLOSE)
-        println(up.append("\n").append(down).toString())
-    }
+    fun printMap(blocks: List<PlayerBlock>) = println(MapMaker.makeMap(blocks))
 
-    private fun appendBlock(
-        playerBlock: PlayerBlock,
-        up: StringBuilder,
-        down: StringBuilder
-    ) {
-        when (playerBlock.getDirection()) {
-            Direction.UP -> {
-                up.append(" $playerBlock ")
-                down.append("   ")
-            }
-            Direction.DOWN -> {
-                up.append("   ")
-                down.append(" $playerBlock ")
-            }
-        }
-    }
-
-    private fun appendMark(up: StringBuilder, down: StringBuilder, mapMark: MapMark) {
-        up.append(mapMark.mark())
-        down.append(mapMark.mark())
-    }
 
     /**
      * 게임의 최종 결과를 정해진 형식에 맞춰 출력한다.
